Compile (mostly) clean with clang-3.8 and gcc-5.4

+++ challenge-0015.cpp
@@ -6,6 +6,6 @@ int main()
     std::string const messages[] = {"BigEndian\n", "LittleEndian\n"};
     auto test = 1;
     auto * p = reinterpret_cast<char *>(&test);
-    std::cout << messages[*p];
+    std::cout << messages[static_cast<int>(*p)];
     return 0;
 }
